Splitted service from metadata
