+++ Files/powerGatePlugin/ErpServices/ErpServices.csproj
@@ -59,17 +59,21 @@
     <Reference Include="System.ServiceModel.Web" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Metadata\Material.cs" />
+    <Compile Include="Metadata\BomHeader.cs" />
+    <Compile Include="Metadata\BomRow.cs" />
     <Compile Include="ErpManager\Implementation\ErpBomManager.cs" />
     <Compile Include="ErpManager\Implementation\ErpDocumentManager.cs" />
     <Compile Include="ErpManager\Interfaces\ErpLogin.cs" />
     <Compile Include="ErpManager\Implementation\ErpItemManager.cs" />
     <Compile Include="ErpManager\Implementation\ErpManager.cs" />
     <Compile Include="ErpManager\Interfaces\IErpManager.cs" />
-    <Compile Include="BomRows.cs" />
-    <Compile Include="BomHeaders.cs" />
-    <Compile Include="Document.cs" />
+    <Compile Include="Services\BomRows.cs" />
+    <Compile Include="Services\BomHeaders.cs" />
+    <Compile Include="Metadata\Document.cs" />
     <Compile Include="ErpManager\Interfaces\ErpMaterialSearchSettings.cs" />
-    <Compile Include="Materials.cs" />
+    <Compile Include="Services\Documents.cs" />
+    <Compile Include="Services\Materials.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WebService.cs" />
   </ItemGroup>
