+++ Files/powerGatePlugin/ErpServices/Metadata/BomHeader.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Data.Services.Common;
+using LiteDB;
+
+namespace ErpServices.Metadata
+{
+    [DataServiceKey("Number")]
+    [DataServiceEntity]
+    public class BomHeader
+    {
+        public string Number { get; set; }
+        [BsonIgnore]
+        public string Description { get; set; }
+        public string State { get; set; }
+        public string UnitOfMeasure { get; set; }
+        public DateTime ModifiedDate { get; set; }
+
+        public List<BomRow> BomRows { get; set;  }
+
+        public BomHeader()
+        {
+            BomRows = new List<BomRow>();
+        }
+    }
+}
