+++ Files/powerGatePlugin/ErpServices/Services/Materials.cs
@@ -1,39 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data.Services.Common;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using ErpServices.Metadata;
 using LiteDB;
 using log4net;
 using powerGateServer.SDK;
 
-namespace ErpServices
+namespace ErpServices.Services
 {
-    [DataServiceKey("Number")]
-    [DataServiceEntity]
-    public class Material
-    {
-        public string Number { get; set; }
-        public string Description { get; set; }
-        public DateTime ModifiedDate { get; set; }
-        public string UnitOfMeasure { get; set; }
-        public string Type { get; set; }
-        public bool IsBlocked { get; set; }
-        public string Category { get; set; }
-        public string Shelf { get; set; }
-        public double Weight { get; set; }
-        public string Dimensions { get; set; }
-        [BsonIgnore]
-        public bool IsVendorSpecified {
-            get => !string.IsNullOrEmpty(VendorNumber);
-            set => _ = value;
-        }
-        public string VendorNumber { get; set; }
-        public string VendorName { get; set; }
-        public string VendorItemNumber { get; set; }
-        public decimal Cost { get; set; }
-    }
-
     public class Materials : ServiceMethod<Material>
     {
         static readonly ILog Log =
