+++ Files/powerGatePlugin/ErpServices/WebService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Configuration;
 using System.Reflection;
+using ErpServices.Services;
 using log4net;
 using powerGateServer.SDK;
